chore: lift signed gas estimate helper into foundry-common
